:tophat: modularize, more test/error work
