+++ test/gist-test.js
@@ -1,14 +1,13 @@
 var vows = require('vows')
 var assert = require('assert')
-
 var Geohub = require('../')
-
-var gist = 6021269
+var gist1 = '6021269'
+var gist2 = '45b401a452cd69e0d5f1'
 
 vows.describe('Gist Access').addBatch({
   'When requesting a gist with geojson': {
     topic: function () {
-      Geohub.gist({ id: gist }, this.callback)
+      Geohub.gist({ id: gist1 }, this.callback)
     },
     'It should return the geojson': function (err, data) {
       assert.equal(err, null)
@@ -19,9 +18,11 @@ vows.describe('Gist Access').addBatch({
       assert.notEqual(data[0].updated_at, null)
     }
   },
+
+  // TODO: the gist for this file does not correspond to the test
   'When requesting a gist with many files': {
     topic: function () {
-      Geohub.gist({ id: '45b401a452cd69e0d5f1' }, this.callback)
+      Geohub.gist({ id: gist2 }, this.callback)
     },
     'It should return the geojson for each file': function (err, data) {
       console.log(err.message)
@@ -36,7 +37,7 @@ vows.describe('Gist Access').addBatch({
   },
   'when getting the sha for a gist': {
     topic: function () {
-      Geohub.gistSha(gist, null, this.callback)
+      Geohub.gistSha(gist1, null, this.callback)
     },
     'It should return a string': function (err, data) {
       assert.equal(err, null)
